requirements.txt
asgiref==3.4.1
Django==4.0
django-stdimage==5.3.0
Pillow==8.4.0
psycopg2==2.9.2
sqlparse==0.4.2
tzdata==2021.5
